+++ apps/erp/app/modules/purchasing/index.ts
@@ -1,4 +1,3 @@
 export * from "./purchasing.models";
 export * from "./purchasing.service";
-export * from "./purchasing.validators";
 export * from "./types";
